Lint
To 'lint' a file is to edit it so that it conforms to style conventions, in
this case those defined by PEP 8. Keeping to convention enables the use
of automated linting software in IDEs and CI/CD that allows developers
to find mistakes early and to identify sections of code that have become
obsolete after a change. (This is all to say, I'm not being a stickler
for 'style' for no reason. Rather, having 'linted' code makes my workflow,
and hopefully everyone else's, much easier.)