Lint; add linter config to ignore long lines
By default, the Flake8 linter tool flags lines over the PEP8 maximum
suggested character length of 79 characters. This is a good suggestion
for new repositories, but in cases of existing code that did not observe
this convention, it's useful to ignore this particular flag (or else the
entire file is highlighted with flags and more meaningful feedback is
hidden among the long-line warnings).