Auto-add ORM objects to session; debug db writes
Jack and I identified an issue where Callicog was not writing to the
database as expected as the experiments proceed. In the process of
debugging, I was puzzled to not see any `db_session.add()` statements
required to add ORM objects to the SQLAlchemy database session--a
prerequisite for these objects to be written to the database on
`commit()`. I added an event listener to ensure that every time an ORM
instance is created, it is immediately added to the session, to avoid
having to go back and identify every single ORM object we created and
ensure it is added to the session. I followed this recipe, provided by
the creator of SQLAlchemy:
https://stackoverflow.com/questions/9873251/how-to-automatically-add-a-sqlalchemy-object-to-the-session

Doing this solved about half of the problem, but for some reason we are
still not seeing any writes to the `protocol_parameter` and `trial_parameter`
database tables.