Move search logic into "schema" code
Instead of creating a stored procedure or a view, I think it's easier to
have the search query logic stored as a query alongside the other
database-related objects. Ultimately the body of the stored procedure is
a query, so it can live as a query alongside the other db business
logic.