Fix TOUCH logic
Logic for TOUCH mode windows was identical to RELEASE mode, probably due
to an old copy-paste error. This is fixed.

NOTE: the release_time variable in this mode will now no longer
represent the "true" release time, but is set to touch_time as a
placeholder. This way the data format will remain a valid datetime and
we will not break downstream processing code. Tracking the true
release_time in TOUCH mode would require a parallel process be executed
to monitor for the release event, complexity and development time which
is not currently warranted given science requirements.