Make IMAGE_DIR relative path
This gets evaluated on the executive computer (Mac) before the task
is pickled and sent to the task agent (Linux). If we use the absolute
path, we end up with something like "/Users/sccn/...", which is not the
correct path on Linux, it would be "/home/sccn/...". Let's see if we can
avoid this just using relative path.